move curated post
